pyproject.toml
[tool.poetry]
name = "fast-api-oauth"
version = "0.1.0"
description = ""
authors = ["Joshua Maldonado"]

[tool.poetry.dependencies]
python = "^3.8"
SQLAlchemy = "^1.4.23"
fastapi = "^0.68.1"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
python-multipart = "^0.0.5"
python-dotenv = "^0.19.0"
psycopg2-binary = "^2.9.1"
alembic = "^1.7.1"
pydantic = {extras = ["email"], version = "^1.8.2"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.dev-dependencies]
black = "^21.9b0"
pytest = "^6.2.5"
requests = "^2.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
